requirements.txt
djangorestframework
Pillow
xrpl-py==2.5.0
drf-yasg